backendOCR/ocr-prescription-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
opencv-python==4.8.1.78
Pillow==10.1.0
numpy>=1.26.0  # Update this line
python-dotenv==1.0.0